still working on the autonomous art
so far we've been able to create a circle but we need to make sure that their is animation when the ball is moving.

Problem: The ball after movinf is still inputed on the screen which means that the line is always there so it like a break along the screen which is not what we are trying to achieve so we need to work on that.